// frontend/src/api/generated/models/GreetingResponse.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Template App API
 * API for the Template Application
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: api-support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GreetingResponse
 */
export interface GreetingResponse {
    /**
     * Technical Unique identifier (TSID). Use for code/logic.
     * @type {number}
     * @memberof GreetingResponse
     */
    id: number;
    /**
     * Functional Business Identifier. Use for display/support.
     * @type {string}
     * @memberof GreetingResponse
     */
    reference: string;
    /**
     * The greeting message content
     * @type {string}
     * @memberof GreetingResponse
     */
    message: string;
    /**
     * Who the greeting is for
     * @type {string}
     * @memberof GreetingResponse
     */
    recipient?: string;
    /**
     * Timestamp when the greeting was created
     * @type {Date}
     * @memberof GreetingResponse
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the GreetingResponse interface.
 */
export function instanceOfGreetingResponse(value: object): value is GreetingResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('reference' in value) || value['reference'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    return true;
}

export function GreetingResponseFromJSON(json: any): GreetingResponse {
    return GreetingResponseFromJSONTyped(json, false);
}

export function GreetingResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GreetingResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'reference': json['reference'],
        'message': json['message'],
        'recipient': json['recipient'] == null ? undefined : json['recipient'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
    };
}

export function GreetingResponseToJSON(json: any): GreetingResponse {
    return GreetingResponseToJSONTyped(json, false);
}

export function GreetingResponseToJSONTyped(value?: GreetingResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'reference': value['reference'],
        'message': value['message'],
        'recipient': value['recipient'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    };
}